+++ tests/testthat/test-mysql.R
@@ -12,7 +12,7 @@ dbExecute(db, "CREATE TABLE orders (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY, 
 dbxInsert(db, "orders", orders[c("city")])
 
 dbExecute(db, "DROP TABLE IF EXISTS events")
-dbExecute(db, "CREATE TABLE events (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY, created_on DATE, updated_at DATETIME(6), deleted_at TIMESTAMP(6))")
+dbExecute(db, "CREATE TABLE events (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY, created_on DATE, updated_at DATETIME(6), deleted_at TIMESTAMP(6), active BOOLEAN)")
 
 test_that("select works", {
   res <- dbxSelect(db, "SELECT * FROM orders ORDER BY id ASC")
@@ -101,6 +101,18 @@ test_that("empty insert works", {
   empty_orders <- data.frame()
   res <- dbxInsert(db, "orders", empty_orders)
   expect_equal(res, empty_orders)
+})
+
+test_that("boolean works", {
+  dbxDelete(db, "events")
+
+  events <- data.frame(active=c(TRUE, FALSE))
+  res <- dbxInsert(db, "events", events)
+
+  expect_equal(res$active, events$active)
+
+  res <- dbxSelect(db, "SELECT * FROM events ORDER BY id")
+  expect_equal(res$active, events$active)
 })
 
 test_that("dates works", {