+++ NEWS.md
@@ -1,5 +1,6 @@
-# dbx 0.1.1
+# dbx 0.1.1 [unreleased]
 
+- Fixed errors with dates and times with SQLite
 - Fixed error with writing dates with MySQL
 - Fixed error with writing times with RPostgreSQL