+++ NEWS.md
@@ -1,8 +1,8 @@
 # dbx 0.1.1 [unreleased]
 
-- Fixed errors with dates and times with SQLite
+- Fixed errors with writing times with PostgreSQL
 - Fixed error with writing dates with MySQL
-- Fixed error with writing times with RPostgreSQL
+- Fixed errors with writing dates and times with SQLite
 
 # dbx 0.1.0