+++ NEWS.md
@@ -1,8 +1,6 @@
 # dbx 0.1.1 [unreleased]
 
-- Fixed errors with writing times with PostgreSQL
-- Fixed error with writing dates with MySQL
-- Fixed errors with writing dates and times with SQLite
+- Fixed errors with dates and times
 
 # dbx 0.1.0