+++ NEWS.md
@@ -1,6 +1,7 @@
 # dbx 0.1.1 [unreleased]
 
 - Fixed errors with dates and times
+- Fixed date and time typecasting with RMySQL
 
 # dbx 0.1.0