+++ NEWS.md
@@ -3,6 +3,7 @@
 - Added support for fractional seconds
 - Fixed errors with dates and times
 - Fixed date and time typecasting with RMySQL
+- Fixed boolean writes with RPostgreSQL
 
 Breaking